Rewrite how visitors work
This is a large-scale refactoring changing how visitors work. The goal
here is to make writing visitors a bit more concise and future-proof.
The major changes here are:

* The `Visitor` trait now has a default implementation for all methods.
  All methods automatically recurse through the AST, by default
  following `Visit`.

* The `Visitor` no longer has an associated `Return` type.

* The `Visit` trait is implemented for a few more types.

* The `Visitor` trait has "base methods" for all sorts of IDs, including
  `FunctionId` and `TableId`. These base methods do no recurse at all,
  but allow hooking references to these types.

* Many visitors have been simplified and/or cleaned up with this new
  visitor pattern.

* The `EmitVisitor` type has been extracted to its own module as it will
  likely grow quite a lot. This module no longer uses the `Visitor`
  trait but instead uses exhaustive matching to ensure we don't forget
  to write down how to emit something.

One of the caveats I ran into was that there's actually cycles in the
wasm AST, notably from branches to their destination blocks. To handle
these cycles and prevent stack overflows this skips them by default when
visiting. Only a DAG-like structure is visited by default.

Currently this doesn't update the dot or display implementations for
functions, but it's hoped that we can start automatically deriving those
soon!