Move the IR display visitor into `#[walrus_expr]`
This should reduce the overhead when adding instructions and also ensure
that all instructions are numbered and such. A sample of the new
debugging output is:

            (func
    (;  0;)   (block
    (;  2;)     (local.set 0
    (;  1;)       (const 9)
                )
    (;  3;)     (loop
    (;  6;)       (br.if (;e0;)
    (;  5;)         (i32.eqz
    (;  4;)           (local.get 0)
                    )
                  )
    (; 10;)       (local.set 0
    (;  9;)         (i32.add
    (;  7;)           (local.get 0)
    (;  8;)           (const 1)
                    )
                  )
                )
    (; 11;)     (const 10)
              )
            )

The tests have been updated for the new output as well. Note that the
expression ids aren't included in the test output but are instead
stripped out, as it's not too critical to be matching them.