Derive `DotVisitor` as part of `#[walrus_expr]`
This commit updates the internal graphviz emission to be auto-generated
as a result of the structural components of `enum Expr`. This is again
aimed at making it easier to add new instructions, ensuring that a
hopefully bare minimum of code needs to be added to support each new
instruction, reducing the amount of boilerplate.