Add dedicated constant value processing
This commit unifies the paths in handling offsets in table elements as
well as initialization expressions in globals. A new `Const` type is
added and a `Const::eval` function is added for transforming an
`InitExpr`.

Notably `Const::Global` exists for non-constant values, and although we
don't parse it just yet support should be plumbed through.