Handle wasm names sections
This commit handles the names sections of wasm executables, all of the
module/function/local names. The actual names of things are currently
stored on the item itself (aka `Option<String>` on `Function` and
`Local`), but I suspect we'll be changing this in the near future.
Shouldn't be too hard to change!