Do not create IR for unreachable expressions
With control frames marked as unreachable, whenever we pop off the operand stack
and the stack is empty, we yield the ID of the expression that made the frame
unreachable. While technically correct, this results in some confusing IR at
times.

This commit makes it so that we don't even construct IR for unreachable
expressions in the first place, since it can't have an effect either way.