Define a walrus-specific hasher for hash maps
One of the hottest parts of profiling `examples/round-trip.rs` on
`clang.wasm` is hashing, presumably with lots of lookups and stores. The
standard library hashing is known to not be optimized for speed by
default, but that's for DoS protection that we're not too concerned
about here.

Instead this commit defines a custom hasher specifically designed to
quickly hash `Id<T>` from the `id-arena` crate. Overally this improves
the runtime of `round-trip` on `clang.wasm` by 15%, reducing the runtime
from 1.9s to 1.6s.