Implement support for all atomic instructions
This commit implements support for all atomic instructions in `walrus`,
adding definitions for all instructions proposed in the threads proposal
as well as hooking up validation to ensure their usage requires shared
memory.

There's not currently a working spec test for these unfortunately, but
this implementation passes against a modified version of the spec tests
currently upstream which wabt presently rejects due to instruction
renaming. We'll be sure to update to run the spec test once it's been
udpated!