Implement some of bulk memory operations
This commit implements the memory-related pieces of the bulk memory
operations proposal. This doesn't currently implement the pieces for
tables because passive table elements depend on `ref.null` and
`ref.func`, both of which are still somewhat up in the air in the
reference types proposal. For now though we can get the benefit of
passive memory initialization as well as memory copies/fills/etc.