Fix actually checking spec test results
Turns out the `spectest-interp` binary will exit with a zero status code
even if some tests fail, notably the execution side of things. This
commit updates our checks of the `spectest-interp` results and actually
looks for `X/Y tests passed.` to handle a case where `X` is not equal to
`Y` and some tests actually failed.