Synthesize `drop` nodes on unreachable
Currently whenever an unreachable instruction is encountered it will
truncate all previous stack operands. If those operands are
side-effect-ful, however, that can be a problem! When an unreachable is
encountered we now synthesize `drop` nodes for all operands on the
stack, ensuring they're all at least executed but their results are
discarded. After all that we push the unreachable instruction into the
block.

This eventually will want to be fixed by us where we don't actually need
to insert `drop` instructions into the final output stream, but for now
it's hopefully good enough.