Add a `ModuleConfig` option for generating debug parser names
This commit starts by growing a `ModuleConfig` structure to place global
configuration for parsing WebAssembly and other sorts of module
configuration. The first option added is one to flag whether or not
debug names should be generated for all IR nodes.

When set, this option will ensure that all functions and locals have a
debug name configured. By default they're all based on the original
indices in the wasm file itself, and then the `name` section, if
present, takes precedent.

This'll also be used by lowering passes so when temporary locals are
created they can be given helpful debugging names as well