Implement automatically updating test assertions
Two main changes here:

* Add `CHECK-ALL:` to exhaustively check the output
* Recognize `WALRUS_BLESS=1` to automatically update all `CHECK-ALL`

The intention here is to make it slightly easier to write tests and
update them over time, where if a sweeping change is made we just run
tests with `WALRUS_BLESS=1` and then review the diff. Hopefully should
make this even easier than it already is to write tests!