Reorganize the crate with less namespaces
This commit reorganizes the crate to dump almost everything into the
global namespace. The `ir` submodule remains, but all other submodules
are now hidden from the public API. All types, however, should be
accessible from either the root of the crate or the `ir` submodule.

This also additionally renames `Const` to `InitExpr` as it's not
actually constant, it's used as an initialization expression.