Don't use stack recursion when parsing instructions
When integrating `walrus` into `wasm-bindgen` everything looked to be
all ship shape except that in debug mode `wasm-bindgen` would very
quickly blow the stack when parsing pretty reasonable wasm files (not
like clang, but like a few test suites). This was due to instruction
validation being stack-recursive and allocating pretty big stack frames
in debug mode.

While the problems didn't show up during release mode at all, I figured
it'd be good to try to tackle this problem to make development of
`wasm-bindgen` a bit easier. To that end this commit updates the
`validate_instruction` function to not recursively call itself, but
rather it's simply called in a loop now. Overall this ended up being a
pretty nice simplification for everything except if/else!