Add an explicit GC pass to walrus
Currently we implicity GC a module when emitting it, but this instead
adds an explicit GC step with the new `delete` functions that we have
now. This way consumers can explicitly choose when to `gc` and then act
on the decisions of the GC to emit different bindings, for example.