fuzz: Use `wasm-opt -ttf` as an additional fuzz test case generator
The `-ttf` flag stands for "translate to fuzz" and instead of treating the input
file as a Wasm binary, it uses it as an stream of "random" values to
deterministically generate Wasm test cases. Larger input files lead to larger
output Wasm test cases.

This commit adapts our existing fuzz harness to use either our existing, simple
`WatGen` generator or use `wasm-opt -ttf` to generate Wasm files that we will
fuzz `walrus` with.