Remove the IR display code and tests
We don't need this code and it will need to be completely reworked with
multi-value, so just get rid of it.

I find the graphviz dot easier to read when debugging the IR (as opposed to the
emitted wasm) anyways...

A few tests only existed in the ir test suite, and I moved them to the
round-trip test suite. Most ir tests had a near exact duplicate in round-trip
already.