fuzz: don't `wasm-opt -ttf --disable-sing-ext`; ignore `wasm-opt` failures instead
`wasm-opt -ttf` will assert that it produces valid wasm modules, but it seems
that it can still generate sign extension instructions even with
`--disable-sign-ext`, which causes validation to fail, and ultimately the
process to exit unsuccessfully. Instead, we just skip and try again when
`wasm-opt` fails, and do the same for if `wat2wasm` cannot interpret what
`wasm-opt` generates (which was why we previously disabled the sign extension
instructions).

See https://github.com/WebAssembly/binaryen/issues/2257