Ensure we don't pre-allocate data segments and then not use them
We were pre-allocating data segments based on the `DataCount` section that we
parsed, but then only re-using those pre-allocated data segments for passive
data segments. For active segments, we created a different structure inside the
module's `Memory`. Then, when we emitted the module again, we would emit
duplicate data segments: both the unused pre-allocated segment and the segment
in the module's `Memory`. Add in the interaction with `data.drop` and which of
the duplicate data segments does it reference, and you get bugs!

Thanks `wasm-opt -ttf` for finding this bug for us :)

Anyways, this commit refactors the way we do data segments (in a breaking way!
we'll need to bump the major version on the next release) so that `ModuleData`
contains all of the data segments, not just the passive segments. And memories
just reference the active data segments that affect them, rather than owning
them. No more duplication and splitting where data segments live depending on
what kind of data segment they are!

Fixes #109