Make `rayon` an optional dependency
This commit makes `rayon` an optional dependency because it is known to
not work on targets such as WebAssembly by default. Additionally it's
not necessarily guaranteed that all users of `walrus` want its parallel
processing capabilties, so this may help it make a lighter dependency in
those situations!

The main use case for me though was I was playing around with some wasm
tooling in Rust compiled to wasm but `walrus` didn't work out-of-the-box
in that context due to its usage of parallel iterators.