Use backwards-compatible encoding of element segments where possible
This is to fix https://github.com/rustwasm/wasm-bindgen/actions/runs/5278136842/jobs/9547894589, which is happening because `@webassemblyjs/wasm-parser` (used by Webpack) doesn't support the newer encoding of element segments that `wasm-encoder` now uses by default.

The alternative to this would be to update `@webassemblyjs/wasm-parser` to support the new encoding, but I think we may as well be backwards-compatible where possible anyway.