feat: upstream ops from WASI-virt
WASI-virt contains functions that are helpful for manipulating modules
and dealing with exports/imports, which would be helpful to an even
wider group if upstreamed here to walrus.

This commit copies and upstreams some operations that were introduced
in WASI-virt for wider use via walrus.

See also: https://github.com/bytecodealliance/WASI-Virt/issues/20

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>